Refactored structs into a directory module.
